WIP on refactor, breaking things apart
