+++ src/pixelator/pna/pixeldataset/precomputed_layouts.py
@@ -1,0 +1,171 @@
+"""Precomputed layouts wrapper for PNA pixel datasets.
+
+Copyright © 2025 Pixelgen Technologies AB.
+"""
+
+from __future__ import annotations
+
+from typing import Iterable
+
+import duckdb
+import numpy as np
+import pandas as pd
+import polars as pl
+
+from pixelator.pna.pixeldataset.io import PixelDataViewer, QueryBuilder
+from pixelator.pna.utils import normalize_input_to_list, normalize_input_to_set
+
+
+class PreComputedLayouts:
+    """Representation of precomputed layouts.
+
+    This contains precomputed layouts for one or more components.
+    """
+
+    def __init__(
+        self,
+        view: PixelDataViewer,
+        components: str | Iterable[str] | None = None,
+        add_marker_counts: bool = True,
+        add_spherical_norm: bool = False,
+    ):
+        """Create a new instance of PreComputedLayouts."""
+        self._view = view
+        self._components = normalize_input_to_set(components)
+        self._add_marker_counts = add_marker_counts
+        self._add_spherical_norm = add_spherical_norm
+        self._query_builder = QueryBuilder()
+
+    @property
+    def components(self) -> set[str]:
+        """Get the component names."""
+        return (
+            self._components
+            if self._components is not None
+            else set(self._view.read_adata().obs.index.to_list())
+        )
+
+    def _pivot_marker_table(self, df: pl.DataFrame) -> pl.DataFrame:
+        """Pivot the joined marker column into marker count columns."""
+        return (
+            df.select(pl.col("*"), val=pl.lit(1))
+            .pivot(
+                on="marker",
+                index=None,
+                values="val",
+                aggregate_function=pl.len().cast(pl.UInt8),
+            )
+            .fill_null(0)
+        )
+
+    def __len__(self) -> int:
+        """Get the nodes in the layouts."""
+        query = self._query_builder.layouts_len_query(
+            normalize_input_to_list(self._components)
+        )
+        with self._view as connection:
+            try:
+                return self._view.execute_scalar(connection, query)
+            except duckdb.CatalogException:
+                return 0
+
+    def is_empty(self) -> bool:
+        """Check if the precomputed layouts are empty."""
+        return len(self) == 0
+
+    def to_df(self) -> pd.DataFrame:
+        """Get the precomputed layouts as a pandas DataFrame."""
+        return self.to_polars().to_pandas()
+
+    def _handle_backwards_comp(self, df: pl.DataFrame) -> pl.DataFrame:
+        # If the norm data is in the output (which it is for some legacy files)
+        # drop them.
+        return df.drop(["x_norm", "y_norm", "z_norm"], strict=False)
+
+    def _post_process(self, df: pl.DataFrame) -> pl.DataFrame:
+        df = self._handle_backwards_comp(df)
+
+        if self._add_spherical_norm:
+            coordinates = df.select(["x", "y", "z"]).to_numpy()
+            normalized_coordinates = pl.DataFrame(
+                coordinates / (1 * np.linalg.norm(coordinates, axis=1))[:, None],
+                schema={
+                    "x_norm": pl.Float32,
+                    "y_norm": pl.Float32,
+                    "z_norm": pl.Float32,
+                },
+            )
+            df = df.hstack(normalized_coordinates)
+
+        return df
+
+    def to_polars(self) -> pl.DataFrame:
+        """Get the precomputed layouts as a polars DataFrame."""
+        query = self._query_builder.layouts_query(
+            components=normalize_input_to_list(self._components),
+            add_marker_counts=self._add_marker_counts,
+        )
+        with self._view as connection:
+            try:
+                if self._add_marker_counts:
+                    layouts = self._view.execute_eager(connection, query)
+                    layouts = self._pivot_marker_table(layouts)
+                    layouts = layouts.drop(["umi", "marker"], strict=False)
+                else:
+                    layouts = self._view.execute_lazy(connection, query)
+                    if isinstance(layouts, pl.LazyFrame):
+                        layouts = layouts.collect()
+            except duckdb.CatalogException:
+                layouts = pl.DataFrame()
+        return self._post_process(layouts)
+
+    def iterator(
+        self, return_polars_df: bool = False
+    ) -> Iterable[tuple[str, pd.DataFrame | pl.DataFrame]]:
+        """Get a stream of layouts.
+
+        Control the trade-off between memory usage and performance by setting
+        the `batch_size` parameter. If you set it to a large number, you will
+        use more memory, but the performance should overall be better.
+
+        :param return_polars_df: If True, return polars DataFrames, otherwise return pandas DataFrames
+
+        :return: A stream of layouts names and associated layout dataframes
+        """
+        with self._view as connection:
+            for component in self.components:
+                query = self._query_builder.layouts_query(
+                    components=[component],
+                    add_marker_counts=self._add_marker_counts,
+                )
+                try:
+                    if self._add_marker_counts:
+                        layouts = self._view.execute_eager(connection, query)
+                        layouts = self._pivot_marker_table(layouts)
+                        layouts = layouts.drop(["umi", "marker"], strict=False)
+                    else:
+                        layouts = self._view.execute_lazy(connection, query)
+                        if isinstance(layouts, pl.LazyFrame):
+                            layouts = layouts.collect()
+                except duckdb.CatalogException:
+                    layouts = pl.DataFrame()
+                component_df = self._post_process(layouts)
+                if return_polars_df:
+                    yield (component, component_df)
+                else:
+                    yield (component, component_df.to_pandas())
+
+    def describe(self) -> str:
+        """Return a description of the PreComputedLayouts."""
+        return f"PreComputedLayouts({len(self.components):,} components, {len(self):,} datapoints)"
+
+    def __str__(self):
+        """Get a string representation of the PreComputedLayouts."""
+        return f"PreComputedLayouts({len(self.components):,} components)"
+
+    def __repr__(self):
+        """Get a string representation of the PreComputedLayouts."""
+        return str(self)
+
+    def _ipython_display_(self):
+        return print(self.describe())
