+++ tests/pna/pixeldataset/test_pna_pixeldataset.py
@@ -1,4 +1,7 @@
-"""Copyright © 2025 Pixelgen Technologies AB."""
+"""Snapshot tests for PNAPixelDataset outputs.
+
+Copyright © 2025 Pixelgen Technologies AB.
+"""
 
 from io import StringIO
 from pathlib import Path
@@ -7,15 +10,10 @@ import duckdb
 import pandas as pd
 import polars as pl
 import pytest
-from polars.testing import assert_frame_equal
 
 from pixelator.common.utils.testing import adata_assert_equal
-from pixelator.pna.pixeldataset import (
-    PixelDatasetConfig,
-    PNAPixelDataset,
-    read,
-)
-from pixelator.pna.pixeldataset.io import PixelDataViewer, PixelFileWriter, PxlFile
+from pixelator.pna.pixeldataset import PixelDatasetConfig, PNAPixelDataset, read
+from pixelator.pna.pixeldataset.io import PixelDataViewer, PxlFile
 from tests.pna.conftest import create_pxl_file
 
 
@@ -76,14 +74,18 @@ class TestPNAPixelDataset:
         adata_assert_equal(adata, adata_data)
 
     def test_adata_adds_transformation_by_default(
-        self, pxl_dataset: PNAPixelDataset, adata_data
+        self,
+        pxl_dataset: PNAPixelDataset,
+        adata_data,  # noqa: ARG002
     ):
         adata = pxl_dataset.adata()
         assert "log1p" in adata.obsm
         assert "clr" in adata.obsm
 
     def test_adata_should_not_mutate_original(
-        self, pxl_dataset: PNAPixelDataset, adata_data
+        self,
+        pxl_dataset: PNAPixelDataset,
+        adata_data,  # noqa: ARG002
     ):
         # Making changes to the adata object should not affect the original
         # i.e. we want to avoid unexpected mutations steming from referencing
@@ -93,42 +95,6 @@ class TestPNAPixelDataset:
 
         assert "new_layer" in adata.layers.keys()
         assert "new_layer" not in pxl_dataset.adata().layers.keys()
-
-    def test_edgelist(self, pxl_dataset: PNAPixelDataset, edgelist_dataframe):
-        edgelist_obj = pxl_dataset.edgelist()
-        edgelist = edgelist_obj.to_polars()
-        assert isinstance(edgelist, pl.DataFrame)
-        edgelist_dataframe = edgelist_dataframe.with_columns(
-            sample=pl.lit("test_sample")
-        )
-        assert_frame_equal(
-            edgelist.sort("component"), edgelist_dataframe.sort("component")
-        )
-
-    def test_proximity(self, snapshot, pxl_dataset):
-        proximity = pxl_dataset.proximity()
-        assert len(proximity) == 3
-
-        proximity_df = proximity.to_polars()
-        assert isinstance(proximity_df, pl.DataFrame)
-        proximity_df = proximity_df.sort("component")
-
-        result = StringIO()
-        proximity_df.write_csv(result)
-        snapshot.assert_match(result.getvalue(), "proximity.csv")
-
-    def test_precomputed_layouts(self, snapshot, pxl_dataset):
-        layouts = pxl_dataset.precomputed_layouts()
-        assert len(layouts) == 34
-        assert len(layouts.components) == 4
-
-        layouts_df = layouts.to_polars()
-        assert isinstance(layouts_df, pl.DataFrame)
-        layouts_df = layouts_df.sort("component").select(sorted(layouts_df.columns))
-
-        result = StringIO()
-        layouts_df.write_csv(result)
-        snapshot.assert_match(result.getvalue(), "layouts.csv")
 
     def test_metadata_returns_expected_mapping(self, pxl_dataset: PNAPixelDataset):
         metadata = pxl_dataset.metadata()
@@ -165,6 +131,8 @@ class TestPNAPixelDataset:
         dataset = PNAPixelDataset.from_pxl_files(pxl_file)
         assert dataset.metadata() == {}
 
+
+class TestPNAPixelDatasetFilter:
     def test_filter_pixel_dataset_by_component(self, pxl_dataset):
         filtered = pxl_dataset.filter(components={"fc07dea9b679aca7"})
         assert filtered.components() == {"fc07dea9b679aca7"}
@@ -315,50 +283,6 @@ class TestPNAPixelDataset:
             pxl_dataset.filter(components=filter_from_polars)
 
 
-class TestPrecomputedLayouts:
-    def test_precomputed_layouts(self, snapshot, pxl_dataset: PNAPixelDataset):
-        layouts = pxl_dataset.precomputed_layouts()
-        assert len(layouts) == 34
-        assert len(layouts.components) == 4
-
-        layouts_df = layouts.to_polars()
-        layouts_df = layouts_df.sort("component").select(sorted(layouts_df.columns))
-
-        result = StringIO()
-        layouts_df.write_csv(result)
-        snapshot.assert_match(result.getvalue(), "layouts.csv")
-
-        assert isinstance(layouts.to_df(), pd.DataFrame)
-
-        # Check iterator contains dataframes as expected
-        for comp_id, component in layouts.iterator():
-            assert isinstance(comp_id, str)
-            assert isinstance(component, pd.DataFrame)
-
-    def test_precomputed_layouts_with_filter(self, pxl_dataset: PNAPixelDataset):
-        filtered = pxl_dataset.filter(components={"fc07dea9b679aca7"})
-        layouts = filtered.precomputed_layouts()
-        assert len(layouts) == 11
-        assert len(layouts.components) == 1
-
-        layouts_df = layouts.to_polars()
-        assert len(layouts_df) == 11
-        assert isinstance(layouts.to_df(), pd.DataFrame)
-        assert len(layouts.to_df()) == 11
-
-        # Check iterator contains dataframes as expected
-        for comp_id, component in layouts.iterator():
-            assert isinstance(comp_id, str)
-            assert isinstance(component, pd.DataFrame)
-
-    def test_precomputed_layouts_with_norm(self, pxl_dataset: PNAPixelDataset):
-        result = pxl_dataset.precomputed_layouts(add_spherical_norm=True).to_polars()
-
-        assert "x_norm" in result.columns
-        assert "y_norm" in result.columns
-        assert "z_norm" in result.columns
-
-
 @pytest.fixture(
     name="pxl_dataset_w_sample_names",
     scope="module",
@@ -442,15 +366,6 @@ class TestPixelDatasetNames:
         actual_sample_name = df.obs["sample"].unique()
         assert actual_sample_name.shape[0] == 1
         assert actual_sample_name[0] == sample_name
-
-
-def test_rewriting_anndata(pxl_file_w_sample_names):
-    pxl_file = pxl_file_w_sample_names
-    pxl = PNAPixelDataset.from_pxl_files(pxl_file)
-    adata = pxl.adata()
-
-    with PixelFileWriter(pxl_file) as writer:
-        writer.write_adata(adata)
 
 
 @pytest.fixture(name="multi_sample_dataset", scope="module")
@@ -642,3 +557,55 @@ class TestPNAPixelDatasetFilterBySample:
         ).to_polars()
         assert set(layouts_df["sample"].unique()) == {"sample1"}
         assert set(layouts_df["component"].unique()) == {target_component}
+
+
+class TestPNAPixelDatasetSnapshots:
+    def test_proximity(self, snapshot, pxl_dataset):
+        proximity = pxl_dataset.proximity()
+        assert len(proximity) == 3
+
+        proximity_df = proximity.to_polars()
+        assert isinstance(proximity_df, pl.DataFrame)
+        proximity_df = proximity_df.sort("component")
+
+        result = StringIO()
+        proximity_df.write_csv(result)
+        snapshot.assert_match(result.getvalue(), "proximity.csv")
+
+    def test_precomputed_layouts(self, snapshot, pxl_dataset):
+        layouts = pxl_dataset.precomputed_layouts()
+        assert len(layouts) == 34
+        assert len(layouts.components) == 4
+
+        layouts_df = layouts.to_polars()
+        assert isinstance(layouts_df, pl.DataFrame)
+        layouts_df = layouts_df.sort("component").select(sorted(layouts_df.columns))
+
+        result = StringIO()
+        layouts_df.write_csv(result)
+        snapshot.assert_match(result.getvalue(), "layouts.csv")
+
+
+class TestPrecomputedLayoutsWrapper:
+    def test_precomputed_layouts_with_filter(self, pxl_dataset: PNAPixelDataset):
+        filtered = pxl_dataset.filter(components={"fc07dea9b679aca7"})
+        layouts = filtered.precomputed_layouts()
+        assert len(layouts) == 11
+        assert len(layouts.components) == 1
+
+        layouts_df = layouts.to_polars()
+        assert len(layouts_df) == 11
+        assert isinstance(layouts.to_df(), pd.DataFrame)
+        assert len(layouts.to_df()) == 11
+
+        # Check iterator contains dataframes as expected
+        for comp_id, component in layouts.iterator():
+            assert isinstance(comp_id, str)
+            assert isinstance(component, pd.DataFrame)
+
+    def test_precomputed_layouts_with_norm(self, pxl_dataset: PNAPixelDataset):
+        result = pxl_dataset.precomputed_layouts(add_spherical_norm=True).to_polars()
+
+        assert "x_norm" in result.columns
+        assert "y_norm" in result.columns
+        assert "z_norm" in result.columns
